+++ solagent-plugins/pumpfun/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "solagent-plugin-pumpfun"
-version = "0.1.2"
+version = "0.1.3"
 edition = "2021"
 authors = ["zTgx <beautifularea@gmail.com>"]
 repository = "https://github.com/zTgx/solagent.rs"
